+++ source/main/utils/safeExitWithCode.js
@@ -8,12 +8,8 @@ export const safeExitWithCode = (exitCode: number) => {
   file.level = false;
   // Flush the stream to the log file and exit afterwards.
   // https://nodejs.org/api/stream.html#stream_writable_end_chunk_encoding_callback
-  file.stream.end(
-    `========== Flushing logs and exiting with code ${exitCode} ===========`,
-    'utf8',
-    () => {
-      app.releaseSingleInstanceLock();
-      app.exit(exitCode);
-    }
-  );
+  file.stream.end('', 'utf8', () => {
+    app.releaseSingleInstanceLock();
+    app.exit(exitCode);
+  });
 };